feat: udisitrital/sga_documentacion#80 Migración de lógica separada del api, implementación de librería util_oas, ajuste de nombre de rutas y variables de entorno
