+++ .env
@@ -1,4 +1,4 @@
-API_NAME=sga_mid_practicas_academicas
-SERVICE_NAME=sgamidpracticasacademicas
+API_NAME=sga_practicas_academicas_mid
+SERVICE_NAME=sgapracticasacademicasmid
 API_BASE_DIR=github.com/udistrital
-SGA_MID_PRACTICAS_ACAD_HTTP_PORT=8095
+SGA_PRACTICAS_ACAD_MID_HTTP_PORT=8095
