+++ conf/app.conf
@@ -1,5 +1,5 @@
 appname = sga_mid_practicas_academicas
-httpport = ${SGA_MID_PRACTICAS_ACAD_HTTP_PORT}
+httpport = ${SGA_PRACTICAS_ACAD_MID_HTTP_PORT}
 runmode = dev
 autorender = false
 copyrequestbody = true
