+++ models/alerta_model.go
@@ -2,6 +2,6 @@ package models
 
 type Alert struct {
 	Type string
-	Code string
+	Code int
 	Body interface{}
 }
