+++ routers/router.go
@@ -9,13 +9,17 @@ package routers
 
 import (
 	"github.com/udistrital/sga_mid_practicas_academicas/controllers"
+	"github.com/udistrital/utils_oas/errorhandler"
 
 	"github.com/astaxie/beego"
 )
 
 func init() {
+
+	beego.ErrorController(&errorhandler.ErrorHandlerController{})
+
 	ns := beego.NewNamespace("/v1",
-		beego.NSNamespace("/practicas_academicas",
+		beego.NSNamespace("/practicas-academicas",
 			beego.NSInclude(
 				&controllers.PracticasAcademicasController{},
 			),
